% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{add_two}
\alias{add_two}
\title{Add two numbers}
\usage{
add_two(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
A numeric value.
}
\description{
Add two numbers
}
\examples{
add_two(1, 2)
}
