% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TjssnStat_help.R
\name{TjssnStat_help}
\alias{TjssnStat_help}
\title{打开 TjssnStat 指定 Stat 模块的参数帮助}
\usage{
TjssnStat_help(Stat = NULL, package = "TjSuper")
}
\arguments{
\item{Stat}{字符串。要查看帮助的分析模块名称，例如 \code{"Super.Table"}。
如果为 \code{NULL}、\code{"help"} 或 \code{"?"}，则列出所有已注册模块。}

\item{package}{字符串。包名，默认 \code{"TjSuper"}。}
}
\value{
不可见地返回帮助对象或模块注册表。
}
\description{
\code{TjssnStat_help()} 用于打开不同 \code{Stat} 参数对应的专门帮助文档。
}
\details{
例如：

\preformatted{
TjssnStat_help("Super.Table")
}

会打开 \code{Stat = "Super.Table"} 的参数说明页。

如果不输入 \code{Stat}，则会列出当前已经注册的所有分析模块。
}
\examples{
TjssnStat_help()
\dontrun{
TjssnStat_help("Super.Table")
}

}
